+++ datasets/ky/parliament/ky_parliament.yml
@@ -36,14 +36,6 @@ assertions:
       attribute: schema
       value: Person
 
-  - metric: entity_count
-    comparison: gt
-    threshold: 50
-    action: fail
-    filter:
-      attribute: schema
-      value: Person
-
   # Upper threshold can warn when lower thresholds probably need updating
   - metric: entity_count
     comparison: lt
