+++ zavod/zavod/meta/assertion.py
@@ -16,13 +16,6 @@ class Comparison(Enum):
     LT = "lt"
 
 
-class Action(Enum):
-    WARN = "warn"
-    """Emit a warning-level log message."""
-    FAIL = "fail"
-    """Fail the job and do not complete producing the dataset."""
-
-
 class Assertion(object):
     """Data assertion specification."""
 
@@ -31,13 +24,12 @@ class Assertion(object):
         comparison_ = type_require(registry.string, config.get("comparison"))
         self.comparison = Comparison(comparison_)
         self.threshold = int(type_require(registry.number, config.get("threshold")))
-        action_ = type_require(registry.string, config.get("action"))
-        self.action = Action(action_)
         if self.metric == Metric.ENTITY_COUNT:
             filter = config.get("filter", {})
             self.filter_attribute = type_require(registry.string, filter.get("attribute"))
             self.filter_value = type_require(registry.string, filter.get("value"))
         else:
+            assert "filter" not in config, self.metric
             self.filter_attribute = None
             self.filter_value = None
